refactor: screen폴더 안으로 이동-1
