+++ client/front/lib/login.dart
@@ -5,12 +5,12 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'sign_up/sign_up.dart';
-import 'main/errand_list/errand_list.dart';
+import 'screens/main/errand_list/errand_list.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_svg/svg.dart';
 
-import 'main/write_errand/write_errand.dart';
+import 'screens/main/write_errand/write_errand.dart';
 void _insertOverlay(BuildContext context) {
 
   if (overlayEntry != null) return;
