+++ client/front/lib/screens/main/errand_list/check_errand/fix_errand/fix_errand_widget/fixminimap.dart
