+++ client/front/lib/screens/main/errand_list/check_errand/show_errand/show_errand_widget/stamp/name_length/name_length3.dart
