+++ client/front/lib/screens/main/errand_list/errand_list.dart
@@ -4,15 +4,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:front/status/status_slide/status_runner.dart';
-import 'package:front/main/write_errand/write_errand.dart';
+import 'package:front/screens/main/write_errand/write_errand.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:intl/intl.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'dart:convert';
 import 'check_errand/check_errand.dart';
-import '../../login.dart';
-import '../../status/status_slide/status_client.dart';
+import '../../../login.dart';
+import '../../../status/status_slide/status_client.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
