+++ client/front/lib/status/status_icons/map.dart
@@ -7,8 +7,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import '../../main/errand_list/errand_list.dart';
-import '../../main/write_errand/write_errand.dart';
+import '../../screens/main/errand_list/errand_list.dart';
+import '../../screens/main/write_errand/write_errand.dart';
 final storage = FlutterSecureStorage();
 void _insertOverlay(BuildContext context) {
