+++ client/front/lib/status/status_slide/status_client.dart
@@ -10,7 +10,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:stomp_dart_client/stomp_dart_client.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../main/errand_list/errand_list.dart';
+import '../../screens/main/errand_list/errand_list.dart';
 import '../status_icons/map.dart';
 import '../status_icons/re-show_errand.dart';
